1./stockIn/code/print orderNo传tid 2.入库日志除了打印入库单，追踪，绑定库位没有做，其余都新增了日志上报 3、更改入库单号为入库任务号； 4、更改揽收任务、清点任务显示的数据跟列表不一致问题 5、SUB-59   【PDA】【清点】【清点】清点时带图片，系统无法提交 6、SUB-57   【PDA】【揽收】【揽收容器】通过揽收容器界面进行绑定，清点，拍照等功能，建议自动带上容器号

+++ app/src/main/java/com/pda/birdex/pda/fragments/CountClearFragment.java
@@ -311,7 +311,7 @@ public class CountClearFragment extends BarScanBaseFragment implements View.OnCl
                         if ("success".equals(object.getString("result"))) {
                             //清点日志上报
                             String upc = edt_upc.getText().toString();
-                            int count = Integer.parseInt(edt_count_length.getText().toString());
+                            long count = Long.parseLong(edt_count_length.getText().toString());
                             String ctNo = edt_count_num.getText().toString();
                             MyApplication.loggingUpload.countClearCommit(getActivity(), tag, orderNo, tid, upc,ctNo,count);
                             T.showShort(getContext(), getString(R.string.taking_upload_suc));
